+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 0.18.6 (Unreleased)
 
 - awsx.ecs.Cluster can be created from an existing aws.ecs.Cluster's id.
+- An `awsx.ec2.Vpc` with `assignGeneratedIpv6CidrBlock: true` will now set
+  `assignIpv6AddressOnCreation: true` by default for child subnets.  This can be overridden by
+  setting that value explicitly to `false` with the subnet's args.
 
 ## 0.18.5 (6/12/2019)